requirements.txt
reportlab>=3.6.12
google-generativeai>=0.0.1
requests>=2.31.0
flask>=2.3.3
python-dotenv>=1.0.0